esc env provider: collapse each cloud into a single env_provider_<cloud>_login.go file
