+++ cmd/esc/cli/env_provider_azure_login_test.go
